+++ src/config/directories.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct DirectoryConfig {
+    pub java: String,
+    pub instance: String,
+}