+++ src/config/launch.rs
@@ -1,0 +1,7 @@
+use crate::minecraft::metadata::Features;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct LaunchConfig {
+    pub features: Features,
+}