+++ src/config/user.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct UserConfig {
+    pub username: String,
+    pub uuid: String,
+    pub version: String,
+}