+++ src/minecraft/launch/assets.rs
@@ -1,0 +1,10 @@
+/*
+
+Get asset index from metadata; store in assets/indexes/{id}.json
+json contains a list of objects with their name and hash
+
+Objects are categorised by first two characters of the has, and stored in assets/objects/{first_two_chars} as files named by their hash with no extension
+
+-- Not sure if necessary, but skin assets are stored in same way under assets/skins
+
+*/