+++ src/minecraft/launch/client.rs
@@ -1,0 +1,8 @@
+/*
+
+Download client jar from version metadata: downloads/client
+rename client to 'minecraft-{version_id}-client.jar', store in libraries/com/mojang/minecraft/{version_id}/minecraft-{version_id}-client.jar
+
+this and libraries will become my -cp or classpath when launching the game
+
+*/