+++ src/minecraft/mod.rs
@@ -1,3 +1,5 @@
+#[allow(unused)]
+pub mod auth;
 pub mod launch;
 pub mod manifest;
 pub mod metadata;